+++ messages/de/base.php
@@ -1,4 +1,0 @@
-<?php
-return array (
-    'Verified Account' => 'Verifizierter Account',
-);