+++ themes/FlexTheme/views/like/widgets/views/likeLink.php
@@ -1,34 +1,22 @@
 <?php
 //FlexTheme
 use \humhub\modules\flexTheme\models\Config;
+use humhub\modules\ui\icon\widgets\Icon;
 //FlexTheme end
 use yii\helpers\Html;
 
 humhub\modules\like\assets\LikeAsset::register($this);
 
 //FlexTheme
-// get Settings
-$icon = Config::getSetting('likeIcon');
+// get Icon
+$color = Config::getSetting('likeIconColor');
+$iconEmpty = Icon::get(Config::getSetting('likeIcon'), ['color' => $color]);
+$iconFull = Icon::get(Config::getSetting('likeIconFull'), ['color' => $color]);
+// get style setting
 $style = Config::getSetting('likeLink');
 
-// Like icon
-if ($icon == 'thumbs_up') {
-    $iconEmpty = '<i class="fa fa-thumbs-o-up"></i>';
-    $iconFull = '<i class="fa fa-thumbs-up"></i>';
-} elseif ($icon == 'heart') {
-    $iconEmpty = '<i class="fa fa-heart-o"></i>';
-    $iconFull = '<i class="fa fa-heart"></i>';
-} elseif ($icon == 'star') {
-    $iconEmpty = '<i class="fa fa-star-o"></i>';
-    $iconFull = '<i class="fa fa-star"></i>';
-}
-
 // additional CSS class
-if ($style == 'text') {
-    $likeContainerClass = 'no-icon';
-} else {
-    $likeContainerClass = $icon . '-' . 'container';
-}
+$likeContainerClass = Html::encode($iconEmpty) . '-' . 'container';
 
 // Like and Unlike Link
 if ($style == 'icon') {
@@ -38,10 +26,10 @@ if ($style == 'icon') {
 	$likeText = $iconEmpty . HTML::tag('span', Yii::t('LikeModule.base', 'Like'), ['class' => 'like-label-both']);
 	$unlikeText = $iconFull . HTML::tag('span', Yii::t('LikeModule.base', 'Unlike'), ['class' => 'unlike-label-both']);
 } else {
+    $likeContainerClass = 'no-icon';
     $likeText = Yii::t('LikeModule.base', 'Like');
     $unlikeText = Yii::t('LikeModule.base', 'Unlike');
 }
-
 ?>
 
 <span class="likeLinkContainer <?= $likeContainerClass ?>" id="likeLinkContainer_<?= $id ?>">